Module encoding_index_japanese::jis0212

JIS X 0212.

From the Encoding Standard:

This is the JIS X 0212 standard. It is only used by the euc-jp decoder due to lack of widespread support elsewhere.

Functions

backward
Returns the index pointer for code point code in this index.

forward
Returns the index code point for pointer code in this index.